refactor(provider): move provider class into toplevel dir
